+++ Acr.Ble.Server.Interfaces/AbstractGattServer.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
+using System.Reactive.Linq;
 
 
 namespace Acr.Ble.Server
@@ -15,7 +16,6 @@ namespace Acr.Ble.Server
         {
             this.internalList = new List<IGattService>();
             this.Services = new ReadOnlyCollection<IGattService>(this.internalList);
-
         }
 
 
@@ -40,10 +40,57 @@ namespace Acr.Ble.Server
 
         public IReadOnlyList<IGattService> Services { get; }
 
-
+        public abstract IObservable<bool> WhenRunningChanged();
         public abstract bool IsRunning { get; }
         public abstract void Start(AdvertisementData adData);
         public abstract void Stop();
+
+
+        IObservable<CharacteristicSubscription> chOb;
+        public virtual IObservable<CharacteristicSubscription> WhenAnyCharacteristicSubscriptionChanged()
+        {
+            this.chOb = this.chOb ?? Observable.Create<CharacteristicSubscription>(ob =>
+            {
+                var cleanup = new List<IDisposable>();
+                foreach (var s in this.Services)
+                {
+                    foreach (var ch in s.Characteristics)
+                    {
+                        cleanup.Add(ch.WhenDeviceSubscriptionChanged().Subscribe(x =>
+                        {
+                            ob.OnNext(new CharacteristicSubscription(ch, x.Device, x.IsSubscribed));
+                        }));
+                    }
+                }
+                return () => 
+                {
+                    foreach (var dispose in cleanup)
+                        dispose.Dispose();
+                };
+            })
+            .Publish()
+            .RefCount();
+
+            return this.chOb;
+        }
+
+
+        public virtual IList<IDevice> GetAllSubscribedDevices()
+        {
+            var list = new Dictionary<Guid, IDevice>();
+            foreach (var s in this.Services) 
+            {
+                foreach (var ch in s.Characteristics)
+                {
+                    foreach (var d in ch.SubscribedDevices)
+                    {
+                        if (!list.ContainsKey(d.Uuid))
+                            list.Add(d.Uuid, d);
+                    }
+                }
+            }
+            return list.Values.ToList();
+        }
 
 
         public IGattService AddService(Guid uuid, bool primary)
@@ -70,6 +117,7 @@ namespace Acr.Ble.Server
             this.ClearNative();
             this.internalList.Clear();
         }
+
 
         protected abstract IGattService CreateNative(Guid uuid, bool primary);
         protected abstract void ClearNative();