+++ Acr.Ble.Server.Interfaces/IGattServer.cs
@@ -6,6 +6,7 @@ namespace Acr.Ble.Server
 {
     public interface IGattServer : IDisposable
     {
+        IObservable<bool> WhenRunningChanged();
         bool IsRunning { get; }
         void Start(AdvertisementData adData);
         void Stop();
@@ -15,16 +16,7 @@ namespace Acr.Ble.Server
         void ClearServices();
         IReadOnlyList<IGattService> Services { get; }
 
-
-        //IGattService CreateService();
-        //IGattCharacteristic CreateCharacteristic();
-        //IGattDescriptor CreateDescriptor();
-
-        // IObservable<bool> WhenRunningStateChanged();
-        // IEnumerable<IDevice> GetConnectedDevices();
-        // Observable<IDevice, bool> WhenDeviceConnectedDisconnected()
-        // IObservable<IGattCharacteristic, IReadRequest> WhenReadRequestReceived();
-        // IObservable<IGattCharacteristic, IWriteRequest> WhenWriteRequestReceived();
-        // IObservable<IGattCharacteristic, bool> WhenCharacteristicSubscriptionStateChanged();
+        IObservable<CharacteristicSubscription> WhenAnyCharacteristicSubscriptionChanged();
+        IList<IDevice> GetAllSubscribedDevices();
     }
 }