+++ Acr.Ble.Server.iOS/GattServer.cs
@@ -2,7 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Reactive.Linq;
+using System.Reactive.Subjects;
 using CoreBluetooth;
+using Foundation;
 
 
 namespace Acr.Ble.Server
@@ -11,19 +13,59 @@ namespace Acr.Ble.Server
     {
         readonly IList<IGattService> services = new List<IGattService>();
         readonly CBPeripheralManager manager;
+        readonly Subject<bool> runningSubj;
 
 
         public GattServer(CBPeripheralManager manager)
         {
             this.manager = manager;
+            this.runningSubj = new Subject<bool>();
         }
 
 
         public override bool IsRunning => this.manager.Advertising;
 
 
+        IObservable<bool> runningOb;
+        public override IObservable<bool> WhenRunningChanged()
+        {
+            this.runningOb = this.runningOb ?? Observable.Create<bool>(ob =>
+            {
+                var handler = new EventHandler<NSErrorEventArgs>((sender, args) => 
+                {
+                    if (args.Error == null)
+                    {
+                        ob.OnNext(true);
+                    }
+                    else 
+                    {
+                        ob.OnError(new ArgumentException(args.Error.LocalizedDescription));
+                    }
+                });
+                var sub = this.runningSubj
+                    .AsObservable()
+                    .Subscribe(x => ob.OnNext(false));
+                
+                this.manager.AdvertisingStarted += handler;
+
+                return () =>
+                {
+                    this.manager.AdvertisingStarted -= handler;
+                    sub.Dispose();
+                };
+            })
+            .Publish()
+            .RefCount();
+            
+            return this.runningOb;
+        }
+
+
         public override void Start(AdvertisementData adData)
         {
+            if (this.manager.Advertising)
+                return;
+            
             if (CBPeripheralManager.AuthorizationStatus != CBPeripheralManagerAuthorizationStatus.Authorized)
                 throw new ArgumentException("Permission Denied - " + CBPeripheralManager.AuthorizationStatus);
 
@@ -47,7 +89,7 @@ namespace Acr.Ble.Server
                     .ServiceUuids
                     .Select(x => CBUUID.FromString(x.ToString()))
                     .ToArray()
-                });
+            });
 
             this.services
                 .Cast<GattService>()
@@ -57,14 +99,14 @@ namespace Acr.Ble.Server
                         .Characteristics
                         .OfType<GattCharacteristic>()
                         .Select(y =>
-                    {
-                        y.Native.Descriptors = y
-                            .Descriptors
-                            .OfType<GattDescriptor>()
-                            .Select(z => z.Native)
-                            .ToArray();
-                        return y.Native;
-                    })
+                        {
+                            y.Native.Descriptors = y
+                                .Descriptors
+                                .OfType<GattDescriptor>()
+                                .Select(z => z.Native)
+                                .ToArray();
+                            return y.Native;
+                        })
                         .ToArray();
 
                     return x.Native;
@@ -76,6 +118,9 @@ namespace Acr.Ble.Server
 
         public override void Stop()
         {
+            if (!this.manager.Advertising)
+                return;
+            
             this.manager.RemoveAllServices();
             this.manager.StopAdvertising();
         }