+++ docs/changelog.md
@@ -1,5 +1,12 @@
 # CHANGE LOG
 
+1.0.1
+
+* [feature] New method for getting all subscribed devices from server instance
+* [feature] Ability to observable all characteristics for a subscription
+* [feature] Observe running state
+* [fix][ios] characteristic observable misfiring
+
 1.0.0
 
 Initial Release